requirements.txt
flask
PyMySQL
numpy
tensorflow-cpu
werkzeug
